Updated inhibitor. Fixed embed parsing, migration. Added guild feature.

+++ src/lib/GreetHandler.ts
@@ -1,11 +1,11 @@
 import {
+    Guild,
     GuildMember,
     Message,
     MessageEmbed,
     MessageEmbedOptions,
     MessageOptions,
     StickerResolvable,
-    TextChannel,
 } from "discord.js";
 import { parsePlaceHolders } from "./functions";
 
@@ -15,21 +15,25 @@ type NotNull<T> = Diff<T, null | undefined>
 interface sendMessageData {
     channel: bigint | null,
     embed: string | null,
-    timeout?: number,
+    timeout: number | null,
 }
-
-interface hasMessageData {
-    channel: NotNull<bigint>,
-    embed: NotNull<string>,
-    timeout?: number,
-}
-
 
 export default class GreetHandler {
-    static JSONErrorMessage = (m: Message) => ({ embeds: [new MessageEmbed()
-        .setTitle("Error")
-        .setDescription("Please provide valid json")
-        .withErrorColor(m)] });
+    static JSONErrorMessage = (m: Message) => ({
+        embeds: [new MessageEmbed()
+            .setTitle("Error")
+            .setDescription("Please provide valid json")
+            .withErrorColor(m)],
+    });
+
+    static emptyMessageOptions = (m: Message | Guild) => ({
+        embeds: [new MessageEmbed()
+            .setTitle("No data")
+            .setTitle("No welcome/bye message set.")
+            .withErrorColor(m),
+        ],
+    });
+
     static async handleGreetMessage(guildMember: GuildMember): Promise<Message | void> {
 
         const db = await guildMember.client.orm.guilds.findFirst({ where: { Id: BigInt(guildMember.guild.id) } });
@@ -38,7 +42,7 @@ export default class GreetHandler {
             return GreetHandler.sendWelcomeLeaveMessage({
                 channel: db.WelcomeChannel,
                 embed: db.WelcomeMessage,
-                timeout: db.WelcomeTimeout || undefined,
+                timeout: db.WelcomeTimeout,
             }, guildMember);
         }
     }
@@ -48,30 +52,31 @@ export default class GreetHandler {
         const db = await guildMember.client.orm.guilds.findFirst({ where: { Id: BigInt(guildMember.guild.id) } });
 
         if (db && db.ByeChannel) {
-            return GreetHandler.sendWelcomeLeaveMessage(<hasMessageData>{
+            return GreetHandler.sendWelcomeLeaveMessage({
                 channel: db.ByeChannel,
                 embed: db.ByeMessage,
-                timeout: db.ByeTimeout || undefined,
+                timeout: db.ByeTimeout,
             }, guildMember);
         }
         return;
     }
 
-    static async createAndParseWelcomeLeaveMessage(data: hasMessageData, guildMember: GuildMember): Promise<MessageOptions> {
-        return JSON.parse(await parsePlaceHolders(data.embed, guildMember));
+    static async createAndParseWelcomeLeaveMessage(data: sendMessageData, guildMember: GuildMember): Promise<MessageOptions> {
+        if (!data.embed) return GreetHandler.emptyMessageOptions(guildMember.guild);
+        return JSON.parse(await parsePlaceHolders(JSON.stringify(data.embed), guildMember));
     }
 
     static async sendWelcomeLeaveMessage(data: sendMessageData, guildMember: GuildMember): Promise<Message | void> {
         if (!data.channel || !data.embed) return;
 
         const channel = guildMember.guild.channels.cache.get(String(data.channel))
-          ?? await guildMember.guild.client.channels.fetch(String(data.channel), { cache: true });
+            ?? await guildMember.guild.client.channels.fetch(String(data.channel), { cache: true });
 
-        if (channel && channel.type !== "GUILD_TEXT" && channel.type !== "GUILD_NEWS") return;
+        if (channel && channel.type !== "GUILD_TEXT" && channel.type !== "GUILD_NEWS" || !channel?.isText()) return;
 
-        const parsedMessageOptions = await GreetHandler.createAndParseWelcomeLeaveMessage(<hasMessageData>data, guildMember);
+        const parsedMessageOptions = await GreetHandler.createAndParseWelcomeLeaveMessage(<sendMessageData>data, guildMember);
 
-        return (channel as TextChannel).send(parsedMessageOptions)
+        return channel.send(parsedMessageOptions)
             .then((m) => {
                 if (data.timeout) {
                     setTimeout(() => m.delete(), data.timeout * 1000);
@@ -88,6 +93,7 @@ export class JSONToMessageOptions implements MessageOptions {
         this.content = any.content;
         this.stickers = any.stickers;
     }
+
     embeds: (MessageEmbed | MessageEmbedOptions)[];
     content?: string | null | undefined;
     stickers?: StickerResolvable[] | undefined;
